Module aws_sdk_qldbsession::error

Modules

bad_request_exception

See BadRequestException

capacity_exceeded_exception

See CapacityExceededException

invalid_session_exception

See InvalidSessionException

limit_exceeded_exception

See LimitExceededException

occ_conflict_exception

See OccConflictException

rate_exceeded_exception

See RateExceededException

Structs

BadRequestException

Returned if the request is malformed or contains an error such as an invalid parameter

CapacityExceededException

Returned when the request exceeds the processing capacity of the ledger.

InvalidSessionException

Returned if the session doesn't exist anymore because it timed out or expired.

LimitExceededException

Returned if a resource limit such as number of active sessions is exceeded.

OccConflictException

Returned when a transaction cannot be written to the journal due to a failure in the

RateExceededException

Returned when the rate of requests exceeds the allowed throughput.

SendCommandError

Enums

SendCommandErrorKind