Module aws_sdk_qldbsession::model

Modules

abort_transaction_request

See AbortTransactionRequest

abort_transaction_result

See AbortTransactionResult

commit_transaction_request

See CommitTransactionRequest

commit_transaction_result

See CommitTransactionResult

end_session_request

See EndSessionRequest

end_session_result

See EndSessionResult

execute_statement_request

See ExecuteStatementRequest

execute_statement_result

See ExecuteStatementResult

fetch_page_request

See FetchPageRequest

fetch_page_result

See FetchPageResult

io_usage

See IoUsage

page

See Page

start_session_request

See StartSessionRequest

start_session_result

See StartSessionResult

start_transaction_request

See StartTransactionRequest

start_transaction_result

See StartTransactionResult

timing_information

See TimingInformation

value_holder

See ValueHolder

Structs

AbortTransactionRequest

Contains the details of the transaction to abort.

AbortTransactionResult

Contains the details of the aborted transaction.

CommitTransactionRequest

Contains the details of the transaction to commit.

CommitTransactionResult

Contains the details of the committed transaction.

EndSessionRequest

Specifies a request to end the session.

EndSessionResult

Contains the details of the ended session.

ExecuteStatementRequest

Specifies a request to execute a statement.

ExecuteStatementResult

Contains the details of the executed statement.

FetchPageRequest

Specifies the details of the page to be fetched.

FetchPageResult

Contains the page that was fetched.

IoUsage

Contains I/O usage metrics for a command that was invoked.

Page

Contains details of the fetched page.

StartSessionRequest

Specifies a request to start a new session.

StartSessionResult

Contains the details of the started session.

StartTransactionRequest

Specifies a request to start a transaction.

StartTransactionResult

Contains the details of the started transaction.

TimingInformation

Contains server-side performance information for a command. Amazon QLDB captures timing

ValueHolder

A structure that can contain a value in multiple encoding formats.